Call ChangedThreadService subscribers with threads
Right now, a version of the thread before the label/star change is passed
to subscribers. Since the code in main.es6 expects the current version of
the thread, this breaks the to-do saving functionality. When you add the
label selected by the user to a thread, that thread is removed from the
todo.txt; when you remove the label, the thread is added to the todo.txt.