Wrap the React import incantation into a module
The following code

let React;
try {
  React = require('mailspring-exports').React;
} catch (e) {
  React = require('react');
}

appears in many files so that they work both during automated testing and
in Mailspring. By putting it in a module, we can use

import React from '../../react';

everywhere else, which is much simpler and allows us select which React to
use in just one place instead of multiple.